+++ app/models/static.rb
@@ -11,6 +11,9 @@ class Static < ApplicationRecord
   has_many :static_members, dependent: :destroy
   has_many :characters, through: :static_members
 
+  has_many :static_invites, dependent: :destroy
+  has_many :invited_characters, through: :static_invites, source: :static
+
   friendly_id :slug_candidates, use: :slugged
 
   def normalize_friendly_id(text)