+++ app/models/static_invite.rb
@@ -1,0 +1,7 @@
+# Represents invites to statics
+class StaticInvite < ApplicationRecord
+  enum status: { send: 0, declined: 1, approved: 2 }, _prefix: :status
+
+  belongs_to :static
+  belongs_to :character
+end