+++ spec/factories/static_invites.rb
@@ -1,5 +1,5 @@
 FactoryBot.define do
-  factory :static_member do
+  factory :static_invite do
     association :static, :guild
     association :character
   end