+++ spec/models/static_spec.rb
@@ -4,6 +4,8 @@ RSpec.describe Static, type: :model do
   it { should belong_to :world }
   it { should have_many(:static_members).dependent(:destroy) }
   it { should have_many(:characters).through(:static_members) }
+  it { should have_many(:static_invites).dependent(:destroy) }
+  it { should have_many(:invited_characters).through(:static_invites).source(:static) }
 
   it 'factory should be valid' do
     static = build :static, :guild