+++ spec/support/sphinx.rb
@@ -21,9 +21,4 @@ RSpec.configure do |config|
     # stop Sphinx at the end of the test suite.
     # ThinkingSphinx::Test.start_with_autostop
   end
-
-  # config.before(:each) do
-    # Index data when running an acceptance spec.
-    # index if example.metadata[:js]
-  # end
 end