+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - search engine
 - api endpoint characters#search
 - rendering search results
+- StaticInvite model
 
 ## [0.6.0] - 2019-11-01
 ### Added